docs/domain_knowledge/0.아이디어 기획서등등 정리 문서들/테스트-azure 컴퓨터비전/requirements.txt
azure-cognitiveservices-vision-computervision>=0.9.0
pillow>=9.0.0
python-dotenv>=0.19.0
msrest>=0.6.21
